function flip bits

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -10,8 +10,8 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int diff = n ^ m;
-	unsigned int b = 0;
+	unsigned long int diff = n ^ m;
+	unsigned long int b = 0;
 
 	while (diff != 0)
 	{
